renamed TextProcecingConstnat
