+++ TentakelsAttacking2/Helper/private/TextProcessing.cpp
@@ -5,7 +5,7 @@
 
 #include "TextProcessing.h"
 #include "AppContext.h"
-#include "TextProcecingConstant.hpp"
+#include "HTextProcecingConstant.hpp"
 #include <raylib.h>
 #include <string_view>
