+++ TentakelsAttacking2/Helper/public/HTextProcecingConstant.hpp
